refactor mixin package
