+++ src/main/java/alkalus/main/mixins/late/witchery/ItemGeneral$3$1Mixin.java
@@ -1,4 +1,4 @@
-package alkalus.main.mixins.witchery;
+package alkalus.main.mixins.late.witchery;
 
 import net.minecraft.world.World;
