+++ src/main/java/alkalus/main/mixins/late/witchery/TileEntitySpinningWheelMixin.java
@@ -1,4 +1,4 @@
-package alkalus.main.mixins.witchery;
+package alkalus.main.mixins.late.witchery;
 
 import java.util.ArrayList;
