+++ src/main/resources/mixins.WitcheryExtras.json
@@ -1,11 +1,7 @@
 {
   "required": true,
   "minVersion": "0.8.5-GTNH",
-  "package": "alkalus.main.mixins",
   "refmap": "mixins.WitcheryExtras.refmap.json",
   "target": "@env(DEFAULT)",
-  "compatibilityLevel": "JAVA_8",
-  "mixins": [],
-  "client": [],
-  "server": []
+  "compatibilityLevel": "JAVA_8"
 }
